+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# Revision history for EwwNotificationDaemon
+# Revision history 
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.1.0 -- 2023-08-31
 
-* First version. Released on an unsuspecting world.
+* Fix `max-notifications` config option
+
+## 0.1.0.0 -- 2023-08-28
+
+* Initial release